moved UsersAndGroups to the pages directory
