+++ frontend/src/stories/UsersAndGroups.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import UsersAndGroups from "../components/regions/UsersAndGroups/Presentation";
+import UsersAndGroups from "../components/pages/UsersAndGroups/Presentation";
 import { BrowserRouter as Router } from "react-router-dom";
 
 import teams from "./fixtures/allTeams.json";
